This is the torn down version of sixbacon's file.
The goals of this document are for the user to enter an email address which can then be used to send the image as an email later. I am also not sure how to make the screen centre. The field form for the original button layout is still there blocking the screen from expanding but I'm not sure how to redact it without causing errors.